requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.1
pandas==2.0.3
openpyxl==3.1.2
email-validator==2.0.0
python-multipart==0.0.6
aiofiles==23.2.1